refactor: extract to file
